+++ src/extension.ts
@@ -11,14 +11,16 @@ import {
   createRegisteredDocumentProviderManager,
   RegisteredDocumentContentProvider,
 } from "./registeredDocumentContentProvider";
-import { defaultExtensionContextManager } from "./extensionContextManager";
 import { GitMergetoolReplacement } from "./gitMergetoolReplacement";
 import {
   createReadonlyDocumentProviderManager,
   ReadonlyDocumentProvider,
 } from "./readonlyDocumentProvider";
 import { RegisterableService } from "./registerableService";
-import { SettingsAssistantCreator } from "./settingsAssistant";
+import {
+  OptionChangeProtocolExporter,
+  SettingsAssistantLauncher,
+} from "./settingsAssistant";
 import { GitMergetoolTemporaryFileOpenHandler } from "./gitMergetoolTemporaryFileOpenHandler";
 import { TemporarySettingsManager } from "./temporarySettingsManager";
 import { VSCodeConfigurator } from "./vSCodeConfigurator";
@@ -33,9 +35,8 @@ let extensionAPI: ExtensionAPI | undefined;
 export async function activate(
   context: ExtensionContext
 ): Promise<ExtensionAPI> {
-  defaultExtensionContextManager.value = context;
   extensionAPI = new ExtensionAPI(
-    ...new ExtensionServicesCreator().createServices()
+    ...new ExtensionServicesCreator().createServices(context)
   );
   await extensionAPI.register();
   return extensionAPI;
@@ -73,37 +74,38 @@ export class ExtensionAPI implements RegisterableService {
 
 class ExtensionServicesCreator {
   public createServices(
+    context: ExtensionContext,
     services: Readonly<Partial<ExtensionServices>> = {}
   ): [ExtensionServices, RegisterableService[]] {
     const registrationOrder: RegisterableService[] = [];
 
     const vSCodeConfigurator =
-      services?.vSCodeConfigurator || new VSCodeConfigurator();
+      services?.vSCodeConfigurator ?? new VSCodeConfigurator();
 
     const readonlyDocumentProviderManager =
-      services.readonlyDocumentProviderManager ||
+      services.readonlyDocumentProviderManager ??
       createReadonlyDocumentProviderManager();
     registrationOrder.push(readonlyDocumentProviderManager);
     const readonlyDocumentProvider =
       readonlyDocumentProviderManager.documentProvider;
 
     const registeredDocumentProviderManager =
-      services.registeredDocumentProviderManager ||
+      services.registeredDocumentProviderManager ??
       createRegisteredDocumentProviderManager();
     registrationOrder.push(registeredDocumentProviderManager);
     const registeredDocumentProvider =
       registeredDocumentProviderManager.documentProvider;
 
-    const zoomManager = services.zoomManager || new ZoomManager();
+    const zoomManager = services.zoomManager ?? new ZoomManager();
     registrationOrder.push(zoomManager);
 
     const temporarySettingsManager =
-      services.temporarySettingsManager ||
-      new TemporarySettingsManager(vSCodeConfigurator);
+      services.temporarySettingsManager ??
+      new TemporarySettingsManager(vSCodeConfigurator, context.globalState);
     registrationOrder.push(temporarySettingsManager);
 
     const diffLayouterManager =
-      services.diffLayouterManager ||
+      services.diffLayouterManager ??
       new DiffLayouterManager(
         vSCodeConfigurator,
         zoomManager,
@@ -117,7 +119,7 @@ class ExtensionServicesCreator {
     const manualMergeProcess = new ManualMergeProcess(diffLayouterManager);
 
     const gitMergetoolReplacement =
-      services.gitMergetoolReplacement ||
+      services.gitMergetoolReplacement ??
       new GitMergetoolReplacement(
         registeredDocumentProvider,
         readonlyDocumentProvider,
@@ -128,14 +130,14 @@ class ExtensionServicesCreator {
     registrationOrder.push(gitMergetoolReplacement);
 
     const temporaryFileOpenManager =
-      services.temporaryFileOpenManager ||
+      services.temporaryFileOpenManager ??
       new GitMergetoolTemporaryFileOpenHandler(
         diffLayouterManager,
         readonlyDocumentProvider
       );
 
     const editorOpenManager =
-      services.editorOpenManager ||
+      services.editorOpenManager ??
       new EditorOpenManager([
         {
           handler: gitMergetoolReplacement,
@@ -149,16 +151,27 @@ class ExtensionServicesCreator {
     registrationOrder.push(editorOpenManager);
 
     const arbitraryFilesMerger =
-      services.arbitraryFilesMerger ||
-      new ArbitraryFilesMerger(diffLayouterManager, readonlyDocumentProvider);
+      services.arbitraryFilesMerger ??
+      new ArbitraryFilesMerger(
+        diffLayouterManager,
+        readonlyDocumentProvider,
+        context.workspaceState
+      );
     registrationOrder.push(arbitraryFilesMerger);
 
-    const settingsAssistantCreator =
-      services.settingsAssistantCreator ||
-      new SettingsAssistantCreator(vSCodeConfigurator);
-    registrationOrder.push(settingsAssistantCreator);
+    const optionChangeProtocolExporter =
+      services.optionChangeProtocolExporter ??
+      new OptionChangeProtocolExporter();
 
-    const mergeAborter = services.mergeAborter || new MergeAborter();
+    const settingsAssistantLauncher =
+      services.settingsAssistantLauncher ??
+      new SettingsAssistantLauncher(
+        vSCodeConfigurator,
+        optionChangeProtocolExporter
+      );
+    registrationOrder.push(settingsAssistantLauncher);
+
+    const mergeAborter = services.mergeAborter ?? new MergeAborter();
     registrationOrder.push(mergeAborter);
 
     return [
@@ -168,7 +181,8 @@ class ExtensionServicesCreator {
         gitMergetoolReplacement,
         readonlyDocumentProviderManager,
         registeredDocumentProviderManager,
-        settingsAssistantCreator,
+        settingsAssistantLauncher,
+        optionChangeProtocolExporter,
         temporarySettingsManager,
         vSCodeConfigurator,
         zoomManager,
@@ -192,7 +206,8 @@ export interface ExtensionServices {
   gitMergetoolReplacement: GitMergetoolReplacement;
   diffLayouterManager: DiffLayouterManager;
   arbitraryFilesMerger: ArbitraryFilesMerger;
-  settingsAssistantCreator: SettingsAssistantCreator;
+  settingsAssistantLauncher: SettingsAssistantLauncher;
+  optionChangeProtocolExporter: OptionChangeProtocolExporter;
   temporaryFileOpenManager: GitMergetoolTemporaryFileOpenHandler;
   editorOpenManager: EditorOpenManager;
   commonMergeCommandsManager: CommonMergeCommandsManager;